vanderval/requirements.txt
amqp==5.3.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==2.2.28
djangorestframework==3.12.4
kombu==5.4.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
setuptools==75.6.0
six==1.17.0
sqlparse==0.4.4
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
